+++ OAuth/ResourceOwner/TrelloResourceOwner.php
@@ -1,0 +1,41 @@
+<?php
+
+/*
+ * This file is part of the HWIOAuthBundle package.
+ *
+ * (c) Hardware.Info <opensource@hardware.info>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace HWI\Bundle\OAuthBundle\OAuth\ResourceOwner;
+
+/**
+ * TrelloResourceOwner
+ *
+ * @author Joseph Bielawski <stloyd@gmail.com>
+ */
+class TrelloResourceOwner extends GenericOAuth1ResourceOwner
+{
+    /**
+     * {@inheritDoc}
+     */
+    protected $options = array(
+        'authorization_url' => 'https://trello.com/1/OAuthAuthorizeToken',
+        'request_token_url' => 'https://trello.com/1/OAuthGetRequestToken',
+        'access_token_url'  => 'https://trello.com/1/OAuthGetAccessToken',
+        'infos_url'         => 'https://api.trello.com/1/members/me?fields=username,fullName,avatarSource,email',
+    );
+
+    /**
+     * {@inheritDoc}
+     */
+    protected $paths = array(
+        'identifier'     => '_id',
+        'nickname'       => 'username',
+        'realname'       => 'fullName',
+        'email'          => 'email',
+        'profilepicture' => 'avatarSource',
+    );
+}